fastapi/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-core
langgraph
openai
mongoengine
python-multipart
python-dotenv
pymongo 